+++ api-rs/src/main/java/nl/vpro/api/rs/v2/page/PageRestService.java
@@ -26,7 +26,7 @@ public interface PageRestService {
     @Path("/")
     Result<Page> list(
             @QueryParam("profile") String profile,
-            @QueryParam("offset") @DefaultValue("0") Integer offset,
+            @QueryParam("offset") @DefaultValue("0") Long offset,
             @QueryParam("max") @DefaultValue(Constants.MAX_RESULTS_STRING) Integer max,
             @QueryParam("mock") @DefaultValue("false") boolean mock);
 
@@ -35,7 +35,7 @@ public interface PageRestService {
     SearchResult<Page> find(
         PageForm form,
         @QueryParam("profile") String profile,
-        @QueryParam("offset") @DefaultValue("0") Integer offset,
+        @QueryParam("offset") @DefaultValue("0") Long offset,
         @QueryParam("max") @DefaultValue(Constants.MAX_RESULTS_STRING) Integer max,
         @QueryParam("mock") @DefaultValue("false") boolean mock);
